feat: спринт 13. попытка 1.

+++ src/main/java/ru/practicum/shareit/item/ItemServiceImpl.java
@@ -70,7 +70,7 @@ public class ItemServiceImpl implements ItemService {
     public List<ItemDto> getItems(Integer userId) {
         List<ItemDto> listItemDto = new ArrayList<>();
         for (Item item : itemStorage.getItems()) {
-            if (item.getOwner().getId() == userId) {
+            if (item.getOwner().getId().intValue() == userId) {
                 listItemDto.add(ItemMapper.toItemDto(item));
             }
         }
